Added root check to docker-remove-stopped

+++ docker-remove-stopped.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+
+if [ "$EUID" != 0 ]
+  then echo "Run me as root!"
+  exit
+fi
+
+
 docker rm $(docker ps -a -q)
 echo "Complete."
 echo "Use docker -a - q to view stopped containers."
